Fix BS touch input bug
this bug is bullshit, wasted a solid hour of time :(

so on touch devices, the device number is not zero for some reason, meaning that touch controls did not work. this took most of the time to figure out.

when starting the game, the device *is* num 0, so the game works for one button press, then it doesnt.

to fix, i set the mouse click to listen on all devices.